+++ COT/__init__.py
@@ -33,6 +33,7 @@ Command modules
   COT.add_disk
   COT.add_file
   COT.deploy
+  COT.deploy_esxi
   COT.edit_hardware
   COT.edit_product
   COT.edit_properties