+++ COT/cli.py
@@ -437,7 +437,7 @@ class CLI(UI):
         """
         from COT.add_disk import COTAddDisk
         from COT.add_file import COTAddFile
-        from COT.deploy import COTDeployESXi
+        from COT.deploy_esxi import COTDeployESXi
         from COT.edit_hardware import COTEditHardware
         from COT.edit_product import COTEditProduct
         from COT.edit_properties import COTEditProperties
@@ -543,8 +543,13 @@ class CLI(UI):
                      "Please contact the COT development team."
                      .format(e.args[0]))
         except EnvironmentError as e:
-            print(e.strerror)
-            sys.exit(e.errno)
+            # EnvironmentError may have both or neither of (errno, strerror).
+            if e.errno is not None:
+                print(e.strerror)
+                sys.exit(e.errno)
+            else:
+                print(e.args[0])
+                sys.exit(1)
         except KeyboardInterrupt:
             sys.exit("\nAborted by user.")
         finally: